server/requirements.txt
annotated-types==0.6.0
blinker==1.6.2
blis==0.7.11
boto3==1.28.73
botocore==1.31.73
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cryptography==41.0.4
cymem==2.0.8
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
mysql-connector-python==8.1.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
packaging==23.1
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
preshed==3.0.9
protobuf==4.21.12
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pypdfium2==4.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
s3transfer==0.7.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.3
Werkzeug==3.0.0
zipp==3.17.0
